Migrate old lexer system to new one.
Introduce the new fancy lexer base.
Remove old lexer code.
Alter the directives system.
Only directly represent some types in the lexer, as directives can now use any symbol.
Lexer only directly supports very common and basic token structures.
For now, it's only NAMESPACE, STRING, and CHAR
NAMESPACE is defined as having a starting character that matches Character.isJavaIdentifierStart(), with the rest of the token matching the predicate as defined by Character.isJavaIdentifierPart().
STRING is defined as Java's string literal, as because Java code can be written directly within the Lexer definiton, and creating a token for each character and getting the directives to work out the actual string is annoying.
CHAR is defined as Java's char literals for the same reason as STRING.

As of now, those are the only directly supported token types within the lexer itself, but it does return tokens for every non-matching character, excluding characters that could start a token.
While those are the only supported characters within the Lexer, the lexer does support comments.
Line comments, and nested block comments. This might be changed with a later release, as keeping the differences between Java and LexiJ to a minimum is desirable.